+++ src/Helpers/AdditionalResponseData.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Efabrica\GraphQL\Helpers;
+
+final class AdditionalResponseData
+{
+    public array $data = [];
+
+    public array $debugData = [];
+}