+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Multi schema loader
+- Ability to add additional data to response
+- GraphQL exception that can contain both safe message and debug message
+- Debug mode to WebonyxDriver
+
+### Changed
+- Type with fields will combine fields from callback and predefined fields
+- Webonyx schema transformer clears cache after transforming schema 
+- OrderArgument key changed to 'column' and order to 'sort_order'
 
 ## [0.1.0] - 2022-10-24
 ### Added